feat: move all methods from EventForwarder to JsonTypeEventForwarder and turn the EventForwarder more generic to implement a custom forwarder
